fixed test

+++ modules/web/site/src/test/kotlin/agartha/site/controllers/mocks/MockedPractitionerService.kt
@@ -174,7 +174,7 @@ class MockedPractitionerService : IPractitionerService {
         return false
     }
 
-    override fun payForAddingVirtualSessions(practitioner: PractitionerDBO, numberOfSessions: Long): Boolean {
+    override fun payForAddingVirtualSessions(practitioner: PractitionerDBO, numberOfSessions: Int): Boolean {
         val practitionerId: String = practitioner._id ?: ""
         val pointsToPay = Settings.COST_ADD_VIRTUAL_SESSION_POINTS * numberOfSessions
         val spiritBankLog = practitioner.spiritBankLog.toMutableList()
